brickserver_playground/requirements.txt
aiofiles
aiosparql
arrow
asyncpg
authlib
docker
fastapi-utils
git+https://github.com/jbkoh/brick_data.git@master
git+https://github.com/jbkoh/fastapi.git@fix-bodyparsing
grpcio
httpx==0.12.0
itsdangerous
jedi
mongoengine
numpy
pandas
protobuf
psycopg2-binary
pydantic
PyJWT
pyparsing==2.3.1
pytest
pytest-env
pytest-ordering
PyYAML
rdflib
redis
requests
starlette
uvicorn
Werkzeug
